feat(data): Add an options to control mapping types.
Added visible as a flag to indicate if this mapping type is externally visible to other systems, or just for internal api use only.
Added editable as a flag to indicate if this mapping type is editable or not, even though it is possible to set editable to true while visible to false, this will have no effect, since no external system could ever change a mapping type that is not visible.